FDG-5404 Move chart into separate file
